cpg_infra/billing_aggregator/monthly_aggregate/requirements.txt
flask
pandas
db-dtypes==1.0.5
google-cloud-bigquery==3.6.0
google-cloud-bigquery-storage==2.19.0
google-api-python-client==2.80.0
